riven update ^^

+++ ProjectRiven/ProjectRiven/DamageHandler.cs
@@ -5,17 +5,17 @@ namespace ProjectRiven
 {
     internal class DamageHandler
     {
-        public static double QDamage(bool useR = false)
+        public static float QDamage(bool useR = false)
         {
-            return new double[] {10, 30, 50, 70, 90}[Riven.Q.Level - 1] +
-                   ((Riven.R.IsReady() && useR && !Player.Instance.HasBuff("RivenFengShuiEngine") ? Player.Instance.TotalAttackDamage * 1.2 : Player.Instance.TotalAttackDamage) /100)*
-                   new double[] {40, 45, 50, 55, 60}[Riven.Q.Level - 1];
+            return (float) (new double[] {10, 30, 50, 70, 90}[Riven.Q.Level - 1] +
+                            ((Riven.R.IsReady() && useR && !Player.Instance.HasBuff("RivenFengShuiEngine") ? Player.Instance.TotalAttackDamage * 1.2 : Player.Instance.TotalAttackDamage) /100)*
+                            new double[] {40, 45, 50, 55, 60}[Riven.Q.Level - 1]);
         }
 
-        public static double WDamage()
+        public static float WDamage()
         {
-            return new double[] {50, 80, 110, 140, 170}[Riven.W.Level - 1] +
-                   1*ObjectManager.Player.FlatPhysicalDamageMod;
+            return (float) (new double[] {50, 80, 110, 140, 170}[Riven.W.Level - 1] +
+                            1*ObjectManager.Player.FlatPhysicalDamageMod);
         }
 
         public static double PassiveDamage()
@@ -24,13 +24,13 @@ namespace ProjectRiven
                    (ObjectManager.Player.BaseAttackDamage + ObjectManager.Player.FlatPhysicalDamageMod);
         }
 
-        public static double RDamage(Obj_AI_Base target, double healthMod = 0)
+        public static float RDamage(Obj_AI_Base target, double healthMod = 0)
         {
             if (!Riven.R.IsLearned) return 0;
             var hpPercent = (target.Health - healthMod > 0 ? 1 : target.Health - healthMod) /target.MaxHealth;
-            return (new double[] {80, 120, 160}[Riven.R.Level - 1]
-                   + 0.6 * Player.Instance.FlatPhysicalDamageMod) *
-                   (hpPercent < 25 ? 3 : (((100 - hpPercent) * 2.67) / 100) + 1);
+            return (float) ((new double[] {80, 120, 160}[Riven.R.Level - 1]
+                             + 0.6 * Player.Instance.FlatPhysicalDamageMod) *
+                            (hpPercent < 25 ? 3 : (((100 - hpPercent) * 2.67) / 100) + 1));
         }
     }
 }
